src/requirements.txt
aiohttp==3.11.13
cloudpickle==3.1.1
docker_py==1.10.6
Flask==3.1.0
graphviz==0.20.3
redis==5.2.1
setuptools==76.0.0
